oe-buildenv-internal: Fix finding build directory
The intent of the env setup scripts is to set BBPATH to point at the
build directory if not already set. This means if the user changes
directory, bitbake can still find the original build directory. The
default bblayers.conf files reset BBPATH to the correct components so
this is safe and restores the behaviour the script was intended to have.

[YOCTO #12163]

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>